YAAUM-36 Add button and other stuff
