+++ common/core/src/main/kotlin/dev/yaaum/common/core/ext/Int.kt
@@ -1,0 +1,3 @@
+package dev.yaaum.common.core.ext
+
+infix fun Int.has(bit: Int) = this.and(bit) != 0
