+++ presentation/core/ui/src/main/kotlin/dev/yaaum/presentation/core/ui/composable/button/base/YaaumButtonAnimation.kt
@@ -1,0 +1,10 @@
+package dev.yaaum.presentation.core.ui.composable.button.base
+
+import androidx.compose.animation.core.Easing
+import androidx.compose.runtime.Immutable
+
+@Immutable
+interface YaaumButtonAnimation {
+    val duration: Int
+    val easing: Easing
+}
