+++ presentation/core/ui/src/main/kotlin/dev/yaaum/presentation/core/ui/composable/button/base/YaaumButtonColors.kt
@@ -1,0 +1,20 @@
+package dev.yaaum.presentation.core.ui.composable.button.base
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Stable
+import androidx.compose.runtime.State
+import androidx.compose.ui.graphics.Color
+
+interface YaaumButtonColors {
+    @Stable
+    @Composable
+    fun borderColor(interactionState: Int, enabled: Boolean): State<Color>
+
+    @Stable
+    @Composable
+    fun foregroundColor(interactionState: Int, enabled: Boolean): State<Color>
+
+    @Stable
+    @Composable
+    fun backgroundColor(interactionState: Int, enabled: Boolean): State<Color>
+}
