+++ presentation/core/ui/src/main/kotlin/dev/yaaum/presentation/core/ui/composable/button/base/YaaumButtonDefaults.kt
@@ -1,0 +1,86 @@
+package dev.yaaum.presentation.core.ui.composable.button.base
+
+import androidx.compose.animation.core.EaseInCirc
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.shape.CutCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.State
+import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import dev.yaaum.common.core.ext.has
+
+@Suppress("MagicNumber")
+object YaaumButtonDefaults {
+    val colors = object : YaaumButtonColors {
+        @Composable
+        override fun borderColor(interactionState: Int, enabled: Boolean): State<Color> {
+            return rememberUpdatedState(
+                when {
+                    !enabled -> YaaumButtonForegroundDisabled
+                    interactionState has YaaumButtonState.FOCUSED -> YaaumButtonBorderFocused
+                    interactionState has YaaumButtonState.HOVER -> YaaumButtonForegroundHovered
+                    else -> YaaumYaaumYaaumButtonForegroundNormal
+                },
+            )
+        }
+
+        @Composable
+        override fun foregroundColor(interactionState: Int, enabled: Boolean): State<Color> {
+            return rememberUpdatedState(
+                when {
+                    !enabled -> YaaumButtonForegroundDisabled
+                    interactionState has YaaumButtonState.HOVER -> YaaumButtonForegroundHovered
+                    else -> YaaumButtonForegroundNormal
+                },
+            )
+        }
+
+        @Composable
+        override fun backgroundColor(interactionState: Int, enabled: Boolean): State<Color> {
+            return rememberUpdatedState(
+                when {
+                    !enabled -> YaaumButtonBackgroundDisabled
+                    interactionState has YaaumButtonState.PRESSED -> YaaumButtonBackgroundPressed
+                    interactionState has YaaumButtonState.HOVER -> YaaumButtonBackgroundHovered
+                    else -> YaaumButtonBackgroundNormal
+                },
+            )
+        }
+    }
+
+    val sizes = object : YaaumButtonSizes {
+        override val iconSize = 32.dp
+        override val borderSize = 3.dp
+        override val contentPadding = PaddingValues(all = 16.dp)
+        override val spacing = 8.dp
+        override val minWidth = 60.dp
+        override val minHeight = 48.dp
+    }
+
+    val animation = object : YaaumButtonAnimation {
+        override val duration = 250
+        override val easing = EaseInCirc
+    }
+
+    val shape = CutCornerShape(topEndPercent = 30, bottomStartPercent = 30)
+
+    val textStyle = TextStyle(
+//        fontFamily = MontserratFont,
+        fontWeight = FontWeight.SemiBold,
+        fontSize = 24.sp,
+    )
+
+    val YaaumButtonBorderFocused = Color(0xFF00FFFF)
+    val YaaumYaaumYaaumButtonForegroundNormal = Color(0xFF1C4587)
+    val YaaumButtonForegroundNormal = Color.White
+    val YaaumButtonForegroundHovered = Color(0xFF1C4587)
+    val YaaumButtonForegroundDisabled = Color(0xFF666666)
+    val YaaumButtonBackgroundNormal = Color(0XFF3C78D8)
+    val YaaumButtonBackgroundHovered = Color(0xFFC9DAF8)
+    val YaaumButtonBackgroundPressed = Color(0xFF1C4587)
+    val YaaumButtonBackgroundDisabled = Color(0xFFEFEFEF)
+}
