+++ presentation/core/ui/src/main/kotlin/dev/yaaum/presentation/core/ui/composable/button/base/YaaumButtonState.kt
@@ -1,0 +1,12 @@
+package dev.yaaum.presentation.core.ui.composable.button.base
+
+object YaaumButtonState {
+    @JvmStatic
+    val HOVER = 1.shl(0)
+
+    @JvmStatic
+    val PRESSED = 1.shl(1)
+
+    @JvmStatic
+    val FOCUSED = 1.shl(2)
+}
